+++ packages/nimble-components/src/table/tests/types.spec.ts
@@ -1,0 +1,9 @@
+import type { TableColumnSortDirection } from '../types';
+
+describe('Table type', () => {
+    it('TableColumnSortDirection fails compile if assigning arbitrary string values', () => {
+        // @ts-expect-error This expect will fail if the enum-like type is missing "as const"
+        const sortDirection: TableColumnSortDirection = 'hello';
+        expect(sortDirection!).toEqual('hello');
+    });
+});