+++ packages/nimble-components/CHANGELOG.md
@@ -1,8 +1,16 @@
 # Change Log - @ni/nimble-components
 
-This log was last generated on Wed, 01 Mar 2023 18:28:57 GMT and should not be manually modified.
+This log was last generated on Thu, 02 Mar 2023 17:48:41 GMT and should not be manually modified.
 
 <!-- Start content -->
+
+## 18.5.8
+
+Thu, 02 Mar 2023 17:48:41 GMT
+
+### Patches
+
+- Add programmatic column sorting support to the table ([ni/nimble@1a1e630](https://github.com/ni/nimble/commit/1a1e630c2741fd00d73f6b6be700aada832143c4))
 
 ## 18.5.7